feat(#12,#13): Conflict/Settings IPC + redaction + route migration + DoH multi-provider
Medium-priority Issue #12 follow-ups and #13 GUI extensions:

* Redaction helpers (S20/S24)
  - types::redact_path(root, path) → project-relative / file_name / <redacted>
  - types::truncate_peer_id for log output
  - PublishUpdate error paths now redact absolute paths
  - presence::unregister_self uses peer_id.short()
  - Chain / Gossip signature-verify errors moved to SynergosNetError::Identity

* GUI Conflicts tab (#13)
  - New IPC: ConflictList + ConflictResolve
  - Conflicts pane renders real data with [Keep Local] / [Accept Remote] / [Manual]
  - CoreConnection::refresh_conflicts + resolve_conflict helpers

* GUI Settings tab (#13)
  - New IPC: ConfigUpdate (mesh_n / max_concurrent_streams / tunnel_hostname)
  - Settings pane has Network Tuning form with Apply button
  - ConnectionCache carries conflicts + feedback

* Route migration periodic driver
  - daemon.rs spawn_periodic_cleanup now calls conduit.try_migrate_route
    for each active QUIC peer every 60s

* DoH hardening (S23)
  - resolve_doh tries Cloudflare then Google, with ResolverOpts.validate = true
  - Single-provider failure no longer takes down name resolution

Tests: 95 passed across workspace (no regressions).

Refs: #12, #13

+++ synergos-core/src/presence/mod.rs
@@ -259,7 +259,7 @@ impl NodeRegistry for PresenceService {
     async fn unregister_self(&self) -> Result<(), NodeRegistryError> {
         let mut local = self.local_node.write().await;
         if let Some(reg) = local.take() {
-            tracing::info!("Unregistering self (peer_id={})", reg.peer_id);
+            tracing::info!("Unregistering self (peer_id={})", reg.peer_id.short());
             self.nodes.remove(&reg.peer_id);
 
             // Gossipsub でオフラインステータスをブロードキャスト
